Extract FirefoxAccount's Profile methods in their own file
