+++ components/fxa-client/src/profile.rs
@@ -1,0 +1,43 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+pub use crate::http_client::ProfileResponse as Profile;
+use crate::{errors::*, http_client::Client, scopes, util, CachedResponse, FirefoxAccount};
+
+// A cached profile response is considered fresh for `PROFILE_FRESHNESS_THRESHOLD` ms.
+const PROFILE_FRESHNESS_THRESHOLD: u64 = 120000; // 2 minutes
+
+impl FirefoxAccount {
+    pub fn get_profile(&mut self, ignore_cache: bool) -> Result<Profile> {
+        let profile_access_token = self.get_access_token(scopes::PROFILE)?.token;
+        let mut etag = None;
+        if let Some(ref cached_profile) = self.profile_cache {
+            if !ignore_cache && util::now() < cached_profile.cached_at + PROFILE_FRESHNESS_THRESHOLD
+            {
+                return Ok(cached_profile.response.clone());
+            }
+            etag = Some(cached_profile.etag.clone());
+        }
+        let client = Client::new(&self.state.config);
+        match client.profile(&profile_access_token, etag)? {
+            Some(response_and_etag) => {
+                if let Some(etag) = response_and_etag.etag {
+                    self.profile_cache = Some(CachedResponse {
+                        response: response_and_etag.response.clone(),
+                        cached_at: util::now(),
+                        etag,
+                    });
+                }
+                Ok(response_and_etag.response)
+            }
+            None => match self.profile_cache {
+                Some(ref cached_profile) => Ok(cached_profile.response.clone()),
+                None => Err(ErrorKind::UnrecoverableServerError(
+                    "Got a 304 without having sent an eTag.",
+                )
+                .into()),
+            },
+        }
+    }
+}
